+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## UNRELEASED
 
 - Added `RestMgmtInterface` endpoint
+- BREAKING CHANGE: Removed fixed setting `AAAPolicy` from `InteropService`
 
 ## v0.15.0